fix: keep proxy generation in build config
